+++ R/summary_lmmSeq.R
@@ -1,4 +1,19 @@
 
+#' Summarise a 'glmmSeq'/'lmmSeq' object
+#' 
+#' Summarise results from [glmmSeq] or [lmmSeq] analysis
+#' 
+#' @param object an object of class `"GlmmSeq"` or `"lmmSeq"`
+#' @param gene an optional character value specifying a single gene whose
+#'   results are summarised
+#' @param digits integer, used for number formatting
+#' @param ... arguments to be passed to other methods
+#' @return
+#' If `gene=NULL` a dataframe of results for all genes is returned. Otherwise
+#' the output of GLMM or LMM model results for a single gene including
+#' coefficients, test statistics, p-values is printed and the dataframe for all
+#' genes is returned invisibly.
+#' @seealso [glmmSeq()], [lmmSeq()]
 #' @export
 
 summary.lmmSeq <- function(object,
@@ -15,7 +30,7 @@ summary.lmmSeq <- function(object,
     })
     do.call(cbind, gp)
   } else {
-    out <- lapply(object@stats, function(i) i[gene,])
+    out <- lapply(object@stats, function(i) i[gene, ])
     if (is(object, "GlmmSeq")) {
       cat("Generalised linear mixed model\n")
       cat(paste0("Method: ", object@info$method, "\n"))
@@ -27,6 +42,8 @@ summary.lmmSeq <- function(object,
     } else {
       cat("Linear mixed model\n")
     }
+    cat("Formula: ")
+    print(object@formula)
     print(out$res)
     cat("\nFixed effects:\n")
     cfdf <- data.frame(Estimate = out$coef,
@@ -34,24 +51,32 @@ summary.lmmSeq <- function(object,
     print(cfdf, digits = digits)
     if (object@info$test.stat == "Wald") {
       cat("\nAnalysis of Deviance Table (Type II Wald chisquare tests)\n")
-      testdf <- data.frame(Chisq = out$Chisq,
-                           Df = out$Df,
-                           `Pr(>Chisq)` = out$pvals, check.names = FALSE)
+      testdf <- data.frame(Chisq = out$Chisq, Df = out$Df,
+                           `Pr(>Chisq)` = out$pvals,
+                           row.names = colnames(object@stats$Chisq),
+                           check.names = FALSE)
+    } else if (object@info$test.stat == "LRT") {
+      cat("\nLikelihood ratio test\nReduced formula: ")
+      print(object@reduced)
+      testdf <- data.frame(Chisq = out$Chisq, Df = out$Df,
+                           `Pr(>Chisq)` = out$pvals,
+                           row.names = " ",
+                           check.names = FALSE)
     } else {
       cat("\nType III Analysis of Variance Table with Satterthwaite's method\n")
-      testdf <- data.frame(NumDF = out$NumDF,
-                           DenDF = out$DenDF,
-                           `F value` = out$Fval,
-                           `Pr(>F)` = out$pvals, check.names = FALSE)
+      testdf <- data.frame(NumDF = out$NumDF, DenDF = out$DenDF,
+                           `F value` = out$Fval, `Pr(>F)` = out$pvals,
+                           row.names = colnames(object@stats$Fval),
+                           check.names = FALSE)
     }
     print(testdf, digits = digits)
     invisible(out)
   }
 }
 
-
+#' @rdname summary.lmmSeq
 #' @export
 
-summary.GlmmSeq <- function(object, ...) {
-  summary.lmmSeq(object, ...)
+summary.GlmmSeq <- function(object, gene = NULL, ...) {
+  summary.lmmSeq(object, gene, ...)
 }